+++ server/AgentDeploy.Tests/Unit/ScriptReaderTests.cs
@@ -20,7 +20,6 @@ namespace AgentDeploy.Tests.Unit
         [Test]
         public async Task TestScriptReader()
         {
-            var operationContext = new OperationContext { OperationCancelled = CancellationToken.None };
             var directoryOptions = new DirectoryOptions { Scripts = "test", Tokens = "test" };
             var deserializer = new DeserializerBuilder()
                 .WithTypeConverter(new YamlStringEnumConverter())
@@ -31,9 +30,9 @@ namespace AgentDeploy.Tests.Unit
             fileReaderMock.Setup(s => s.ReadAsync(It.IsAny<string?>(), It.IsAny<CancellationToken>())).ReturnsAsync("command: echo test");
             var scriptReaderLoggerMock = new Mock<ILogger<ScriptReader>>();
             
-            var scriptReader = new ScriptReader(operationContext, deserializer, fileReaderMock.Object, directoryOptions, scriptReaderLoggerMock.Object);
+            var scriptReader = new ScriptReader(directoryOptions, deserializer, fileReaderMock.Object, scriptReaderLoggerMock.Object);
 
-            var script = await scriptReader.Load("test");
+            var script = await scriptReader.Load("test", It.IsAny<CancellationToken>());
             
             Assert.NotNull(script);
             Assert.AreEqual("test", script!.Name);
@@ -46,7 +45,6 @@ namespace AgentDeploy.Tests.Unit
         [TestCase("none", ConcurrentExecutionLevel.None)]
         public async Task TestScriptReader_ParseConcurrencyLevel(string yamlText, ConcurrentExecutionLevel concurrentExecutionLevel)
         {
-            var operationContext = new OperationContext { OperationCancelled = CancellationToken.None };
             var directoryOptions = new DirectoryOptions { Scripts = "test", Tokens = "test" };
             var deserializer = new DeserializerBuilder()
                 .WithTypeConverter(new YamlStringEnumConverter())
@@ -57,9 +55,9 @@ namespace AgentDeploy.Tests.Unit
             fileReaderMock.Setup(s => s.ReadAsync(It.IsAny<string?>(), It.IsAny<CancellationToken>())).ReturnsAsync($"command: echo test\nconcurrency: {yamlText}");
             var scriptReaderLoggerMock = new Mock<ILogger<ScriptReader>>();
             
-            var scriptReader = new ScriptReader(operationContext, deserializer, fileReaderMock.Object, directoryOptions, scriptReaderLoggerMock.Object);
+            var scriptReader = new ScriptReader(directoryOptions, deserializer, fileReaderMock.Object, scriptReaderLoggerMock.Object);
 
-            var script = await scriptReader.Load("test");
+            var script = await scriptReader.Load("test", It.IsAny<CancellationToken>());
             
             Assert.NotNull(script);
             Assert.AreEqual("test", script!.Name);
@@ -70,7 +68,6 @@ namespace AgentDeploy.Tests.Unit
         [Test]
         public async Task TestScriptReader_Null()
         {
-            var operationContext = new OperationContext { OperationCancelled = CancellationToken.None };
             var directoryOptions = new DirectoryOptions { Scripts = "test", Tokens = "test" };
             var deserializer = new DeserializerBuilder()
                 .WithNamingConvention(UnderscoredNamingConvention.Instance)
@@ -79,9 +76,9 @@ namespace AgentDeploy.Tests.Unit
             fileReaderMock.Setup(s => s.FindFile(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string[]>())).Returns(default(string));
             var scriptReaderLoggerMock = new Mock<ILogger<ScriptReader>>();
             
-            var scriptReader = new ScriptReader(operationContext, deserializer, fileReaderMock.Object, directoryOptions, scriptReaderLoggerMock.Object);
+            var scriptReader = new ScriptReader(directoryOptions, deserializer, fileReaderMock.Object, scriptReaderLoggerMock.Object);
 
-            var script = await scriptReader.Load("test");
+            var script = await scriptReader.Load("test", It.IsAny<CancellationToken>());
             
             Assert.Null(script);
         }
